requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.2.1
djlint==1.36.4
EditorConfig==0.17.0
eval_type_backport==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsbeautifier==1.15.4
json5==0.12.0
mistralai==1.7.1
pathspec==0.12.1
pillow==11.2.1
polib==1.2.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==3.1.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
